+++ lib/creds/railtie.rb
@@ -1,0 +1,7 @@
+module Creds
+  class Railtie < ::Rails::Railtie
+    rake_tasks do
+      Dir[File.join(__dir__, "../tasks/**/*.rake")].each { |f| load f }
+    end
+  end
+end