+++ spec/creds_spec.rb
@@ -9,12 +9,15 @@ RSpec.describe Creds do
       config.logger = Logger.new(STDOUT)
       config.eager_load = false
       config.require_master_key = true
+
+      # Silence warning
+      config.active_support.to_time_preserves_timezone = :zone
     end
 
     RailsTestApp.initialize!
 
     # reset cache
-    Creds.instance.instance_variable_set(:@credentials, nil)
+    Creds.instance_variable_set(:@cache, nil)
 
     write_config({})
   end
@@ -33,65 +36,38 @@ RSpec.describe Creds do
   end
 
   it "returns Rails credentials scoped to env" do
-    write_config(test: {super_secret: "shh!"})
-    expect(Creds.super_secret).to(eq("shh!"))
-  end
-
-  it "nulls out everything when SECRET_KEY_BASE_DUMMY=1" do
-    ENV["SECRET_KEY_BASE_DUMMY"] = "1"
-    write_config(test: {super_secret: "shh!"})
-    expect(Creds.other_secret).to(be_nil)
-  ensure
-    ENV["SECRET_KEY_BASE_DUMMY"] = nil
-  end
-
-  it "merges top-level credentials" do
-    write_config(super_secret: "shh!", test: {other_secret: "SHH!"})
+    write_config(
+      <<-YAML
+      test:
+        super_secret: "shh!"
+      YAML
+    )
     expect(Creds.super_secret).to(eq("shh!"))
   end
 
   it "raises MissingKeyError on missing keys" do
-    write_config(test: {super_secret: "shh!"})
+    write_config(
+      <<-YAML
+      test:
+        super_secret: "shh!"
+      YAML
+    )
     expect { Creds.non_existing_key }.to(raise_error(Creds::MissingKeyError))
   end
 
   it "raises MissingEnvError on missing env" do
-    write_config(development: {})
-    expect { Creds.any_key }.to(raise_error(Creds::MissingEnvError))
+    write_config(
+      <<-YAML
+        development:
+          super_secret: "shh!"
+      YAML
+    )
+    expect { Creds.super_secret }.to(raise_error(Creds::MissingEnvError))
   end
 
-  it "converts to hash" do
-    write_config(test: {super_secret: "shh!"})
-    expect(Creds.to_h).to(eq(super_secret: "shh!"))
-  end
-
-  it "logs and returns null creds when no encrypted credentials" do
-    allow(File).to(
-      receive(:exist?).with(Rails.root.join("config/credentials.yml.enc")) { false }
-    )
-    expect(Rails.logger).to(
-      receive(:warn).with(Creds::MissingCredentialsWarning)
-    )
-
-    result = Creds.some_key
-
-    expect(result).to(be(nil))
-  end
-
-  it "raises with special error when credentials exist but key is missing" do
-    allow(File).to(
-      receive(:exist?).with(Rails.root.join("config/credentials.yml.enc")) { true }
-    )
-    allow(File).to(
-      receive(:exist?).with(Rails.root.join("config/master.key")) { false }
-    )
-
-    expect { Creds.some_key }.to(raise_error(Creds::MissingMasterKeyError))
-  end
-
-  def write_config(conf)
+  def write_config(yaml)
     Rails.application.credentials.change do |path|
-      File.open(path, "w") { |f| f.write(conf.to_yaml) }
+      File.open(path, "w") { |f| f.write(yaml) }
     end
   end
 end